Change some settings
