+++ .idea/libraries/JFX.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="JFX">
+    <CLASSES>
+      <root url="file://$PROJECT_DIR$/../openjfx-19_linux-x64_bin-sdk/javafx-sdk-19/lib" />
+    </CLASSES>
+    <JAVADOC />
+    <NATIVE>
+      <root url="file://$PROJECT_DIR$/../openjfx-19_linux-x64_bin-sdk/javafx-sdk-19/lib" />
+    </NATIVE>
+    <SOURCES />
+    <jarDirectory url="file://$PROJECT_DIR$/../openjfx-19_linux-x64_bin-sdk/javafx-sdk-19/lib" recursive="false" />
+  </library>
+</component>
