--- .idea/libraries/lib.xml
@@ -1,10 +1,0 @@
-<component name="libraryTable">
-  <library name="lib">
-    <CLASSES>
-      <root url="file://$USER_HOME$/Downloads/openjfx-19_windows-x64_bin-sdk/javafx-sdk-19/lib" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-    <jarDirectory url="file://$USER_HOME$/Downloads/openjfx-19_windows-x64_bin-sdk/javafx-sdk-19/lib" recursive="false" />
-  </library>
-</component>
