+++ Bomberman.iml
@@ -10,13 +10,14 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="Maven: org.openjfx:javafx-controls:15" level="project" />
-    <orderEntry type="library" name="Maven: org.openjfx:javafx-controls:win:15" level="project" />
-    <orderEntry type="library" name="Maven: org.openjfx:javafx-fxml:15" level="project" />
-    <orderEntry type="library" name="Maven: org.openjfx:javafx-fxml:win:15" level="project" />
-    <orderEntry type="library" name="Maven: org.openjfx:javafx-base:15" level="project" />
-    <orderEntry type="library" name="Maven: org.openjfx:javafx-base:win:15" level="project" />
-    <orderEntry type="library" name="Maven: org.openjfx:javafx-graphics:15" level="project" />
-    <orderEntry type="library" name="Maven: org.openjfx:javafx-graphics:win:15" level="project" />
+    <orderEntry type="library" name="JFX" level="project" />
+    <orderEntry type="library" name="Maven: org.openjfx:javafx-controls:19" level="project" />
+    <orderEntry type="library" name="Maven: org.openjfx:javafx-controls:win:19" level="project" />
+    <orderEntry type="library" name="Maven: org.openjfx:javafx-fxml:19" level="project" />
+    <orderEntry type="library" name="Maven: org.openjfx:javafx-fxml:win:19" level="project" />
+    <orderEntry type="library" name="Maven: org.openjfx:javafx-base:19" level="project" />
+    <orderEntry type="library" name="Maven: org.openjfx:javafx-base:win:19" level="project" />
+    <orderEntry type="library" name="Maven: org.openjfx:javafx-graphics:19" level="project" />
+    <orderEntry type="library" name="Maven: org.openjfx:javafx-graphics:win:19" level="project" />
   </component>
 </module>
