+++ src/bomberman/entities/Bomber.java
@@ -1,12 +1,12 @@
-package uet.oop.bomberman.entities;
+package bomberman.entities;
 
 import javafx.scene.SnapshotParameters;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.paint.Color;
-import uet.oop.bomberman.MapTiles;
-import uet.oop.bomberman.inPut.handleInput;
+import bomberman.MapTiles;
+import bomberman.inPut.handleInput;
 
 import java.util.ArrayList;
 import java.util.List;
