+++ src/bomberman/entities/Entity.java
@@ -1,11 +1,11 @@
-package uet.oop.bomberman.entities;
+package bomberman.entities;
 
 import javafx.scene.SnapshotParameters;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.paint.Color;
-import uet.oop.bomberman.graphics.Sprite;
+import bomberman.graphics.Sprite;
 
 public abstract class Entity {
     //Tọa độ X tính từ góc trái trên trong Canvas
