+++ src/bomberman/entities/Portal.java
@@ -1,10 +1,10 @@
-package uet.oop.bomberman.entities;
+package bomberman.entities;
 
 import javafx.scene.image.Image;
 
-public class Grass extends Entity {
+public class Portal extends Entity {
 
-    public Grass(int x, int y, Image img) {
+    public Portal(int x, int y, Image img) {
         super(x, y, img);
     }
