--- src/uet/oop/bomberman/inPut/Main.java
@@ -1,11 +1,0 @@
-package uet.oop.bomberman.inPut;
-
-import javafx.application.Application;
-import uet.oop.bomberman.BombermanGame;
-import uet.oop.bomberman.MapTiles;
-
-public class Main {
-    public static void main(String[] args) {
-        Application.launch(BombermanGame.class);
-    }
-}
